alter actix with hyper

+++ src/config/server.rs
@@ -2,7 +2,7 @@ use serde::Serialize;
 
 #[derive(Debug, Serialize, Clone)]
 pub struct Server {
-    pub bind: (String, i32),
+    pub bind: (String, u16),
     #[serde(rename = "pathPrefix")]
     pub path_prefix: Option<String>,
 }
